+++ nas-20170626/alibabacloud_nas20170626/models/_describe_file_systems_response_body.py
@@ -671,9 +671,11 @@ class DescribeFileSystemsResponseBodyFileSystemsFileSystemOptions(DaraModel):
         self,
         enable_abe: bool = None,
         enable_oplock: bool = None,
+        vsc_access_point_access_only: bool = None,
     ):
         self.enable_abe = enable_abe
         self.enable_oplock = enable_oplock
+        self.vsc_access_point_access_only = vsc_access_point_access_only
 
     def validate(self):
         pass
@@ -689,6 +691,9 @@ class DescribeFileSystemsResponseBodyFileSystemsFileSystemOptions(DaraModel):
         if self.enable_oplock is not None:
             result['EnableOplock'] = self.enable_oplock
 
+        if self.vsc_access_point_access_only is not None:
+            result['VscAccessPointAccessOnly'] = self.vsc_access_point_access_only
+
         return result
 
     def from_map(self, m: dict = None):
@@ -698,6 +703,9 @@ class DescribeFileSystemsResponseBodyFileSystemsFileSystemOptions(DaraModel):
 
         if m.get('EnableOplock') is not None:
             self.enable_oplock = m.get('EnableOplock')
+
+        if m.get('VscAccessPointAccessOnly') is not None:
+            self.vsc_access_point_access_only = m.get('VscAccessPointAccessOnly')
 
         return self